app/chunking/pyproject.toml
[project]
name = "chunking"
version = "0.1.0"
description = "Chunks a JSON input"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.54",
    "langchain-text-splitters>=1.1.1",
    "python-dotenv>=1.2.1",
    "utils",
]

[tool.uv.sources]
utils = { path = "../utils", editable = true }
